+++ sapp/decorators.py
@@ -1,23 +1,32 @@
-class WithContext(object):
-    def __init__(self, application, context_name='context', args=None):
-        self.context_name = context_name
+from collections.abc import Iterable
+
+
+class Decorator(object):
+    def __init__(self, application, values=[]):
         self.application = application
-        self.args = args or []
+        self.values = values
+        self.context = None
 
     def __call__(self, fun):
         def wrapper(*args, **kwargs):
-            with self.application as context:
-                return self.fun_with_context(fun, context, args, kwargs)
+            ctx = self.application._enter_context()
+            kwargs = dict(kwargs)
+            try:
+                if self.values == []:
+                    if "ctx" not in kwargs:
+                        kwargs["ctx"] = ctx
+                elif isinstance(self.values, str):
+                    if self.values not in kwargs:
+                        kwargs[self.values] = getattr(ctx, self.values)
+                elif isinstance(self.values, Iterable):
+                    for key in self.values:
+                        if key not in kwargs:
+                            kwargs[key] = getattr(ctx, key)
+                else:
+                    raise AttributeError("Wrong argument type!")
+                return fun(*args, **kwargs)
+            finally:
+                # TODO: we probably need to pass here something
+                self.application._exit_context(None, None, None)
+
         return wrapper
-
-    def fun_with_context(self, fun, context, args, kwargs):
-        if self.args:
-            for arg in self.args:
-                self.set_kwarg(kwargs, arg, getattr(context, arg))
-        else:
-            kwargs[self.context_name] = context
-        return fun(*args, **kwargs)
-
-    def set_kwarg(self, kwargs, arg, value):
-        if arg not in kwargs:
-            kwargs[arg] = value