+++ sapp/tests/test_decorators.py
@@ -4,7 +4,7 @@ from unittest.mock import sentinel
 from pytest import fixture
 
 from sapp.configurator import Configurator
-from sapp.decorators import WithContext
+from sapp.decorators import Decorator
 from sapp.plugin import Plugin
 
 
@@ -28,7 +28,7 @@ class SampleConfigurator(Configurator):
         self.add_plugin(MockedPlugin(['var1', 'var2']))
 
 
-class TestWithContext(object):
+class TestDecorator(object):
     @fixture
     def configurator(self):
         conf = SampleConfigurator()
@@ -37,20 +37,9 @@ class TestWithContext(object):
 
     def test_normal_context(self, configurator):
         """
-        WithContext should append 'context' to the fun's arguments
+        Decorator should append 'ctx' to the fun's arguments
         """
-        @WithContext(configurator)
-        def fun(context):
-            assert context.var1 == configurator.var1
-            assert context.var2 == configurator.var2
-
-        fun()
-
-    def test_with_changed_context_name(self, configurator):
-        """
-        WithContext should append context as a local name fo the fun's arguments
-        """
-        @WithContext(configurator, 'ctx')
+        @Decorator(configurator)
         def fun(ctx):
             assert ctx.var1 == configurator.var1
             assert ctx.var2 == configurator.var2
@@ -59,9 +48,9 @@ class TestWithContext(object):
 
     def test_with_specyfied_values_to_unpack(self, configurator):
         """
-        WithContext should append values from context if specyfied.
+        Decorator should append values from ctx if specyfied.
         """
-        @WithContext(configurator, args=['var1', 'var2'])
+        @Decorator(configurator, ['var1', 'var2'])
         def fun(var1, var2):
             assert var1 == configurator.var1
             assert var2 == configurator.var2
@@ -70,10 +59,10 @@ class TestWithContext(object):
 
     def test_with_values_explicitly_passed(self, configurator):
         """
-        WithContext should not append values from context if those values are
+        Decorator should not append values from ctx if those values are
         passed thru as named argument specyfied.
         """
-        @WithContext(configurator, args=['var1', 'var2'])
+        @Decorator(configurator, ['var1', 'var2'])
         def fun(var1, var2):
             assert var1 == sentinel.var1
             assert var2 == configurator.var2